requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.8
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
env==0.1.0
environs==10.3.0
gunicorn==19.9.0
marshmallow==3.20.2
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==5.1.0
